Basic annotations

+++ 0x00-python_variable_annotations/2-floor.py
@@ -7,4 +7,4 @@ and returns the floor of the float.'''
 
 def floor (n: float) -> int:
     ''' Return floor of n. '''
-    return math.floor(n)
+    return int(math.floor(n))
